% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_missing_data.R
\name{analyze_missing_data}
\alias{analyze_missing_data}
\title{Analyze Missing Data}
\usage{
analyze_missing_data(hapmap_data)
}
\arguments{
\item{hapmap_data}{A data frame containing hapmap data.}
}
\value{
A data frame with an additional column for the percentage of missing data per snp.
}
\description{
Analyze Missing Data
}
