% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_minor_allele_freq.R
\name{calculate_minor_allele_frequency}
\alias{calculate_minor_allele_frequency}
\title{Calculate Minor Allele Frequency}
\usage{
calculate_minor_allele_frequency(hapmap_data)
}
\arguments{
\item{hapmap_data}{A data frame containing hapmap data.}
}
\value{
A data frame with an additional column for Minor Allele Frequency (MAF).
}
\description{
Calculate Minor Allele Frequency
}
