% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_snps.R
\name{filter_snps}
\alias{filter_snps}
\title{Filter SNPs by MAF and Missing Data Thresholds}
\usage{
filter_snps(hapmap_data, maf_threshold = 0.05, missing_threshold = 0.05)
}
\arguments{
\item{hapmap_data}{A data frame containing hapmap data.}

\item{maf_threshold}{The minimum minor allele frequency to retain a SNP.}

\item{missing_threshold}{The maximum percentage of missing data to retain a SNP.}
}
\value{
A filtered data frame of SNPs.
}
\description{
Filter SNPs by MAF and Missing Data Thresholds
}
