% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hapmap.R
\name{read_hapmap}
\alias{read_hapmap}
\title{Read Hapmap File}
\usage{
read_hapmap(file_path, sep = "\\t")
}
\arguments{
\item{file_path}{The file path to the hapmap file.}

\item{sep}{description}
}
\value{
A data frame containing the hapmap data after validation.
}
\description{
Read Hapmap File
}
