% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_hapmap.R
\name{summarize_hapmap}
\alias{summarize_hapmap}
\title{Summarize Hapmap Data}
\usage{
summarize_hapmap(hapmap_data)
}
\arguments{
\item{hapmap_data}{A data frame of hapmap data.}
}
\value{
A list containing summary statistics like the number of SNPs, average MAF, and missing data percentage.
}
\description{
Summarize Hapmap Data
}
